Remove Buffer as a dependency.
