+++ js/hang/src/util/hacks.ts
